requirements.txt
dm-haiku>=0.0.11
flax>=0.7.4
jax>=0.4.2
matplotlib>=3.8.0
numpy>=1.12
optax>=0.1.7
tensorflow>=2.13.0
tensorflow-datasets>=4.9.3
